refactor(Tooltip): Make consistent
Rejigging to be consistent with other components
